Added totInAltWLog and totInAltWAsinh
Logarithm and inverse hyperbolic sine functionality added to totInAltW to handle cases where the dyadic variable used to weight behaviour contributions is highly skewed and/or a variable where change is most meaningfully measured as a percentage change or magnitude change, rather than a unit change.

Previously, I had been taking the logarithm of aid values outside and prior to the RSiena model as a convenience measure, but this distorts the mathematical properties of the change statistic and downweights the relative contribution of large-value donors relative to small-value donors.